Shhh!  Don't tell anyone, but Estlcam 11 integration has been added.  Ctrl + S won't work unless you setup something like AutoHotKey.com app to simulate menu item click when Ctrl + S is pressed.  See https://forum.v1e.com/t/no-undo/52182/58 for details.
Am time constrained, so I've made a best effort attempt here to support EsltCam 11, it seems to work when I did some basic testing.  But it's not a scenario I use, or will be testing going forwards.  Am sharing this to help jump start anyone using EstlCam 11 because they have a bunch of projects in that format still (afaik there's no upgrade conversion tool for migrating .e10 to e12).  Good luck to the brave souls that try this!